Damage Detection using CNN
This project aimed to develop a convolutional neural network (CNN) model for damage detection in images. It utilized a dataset consisting of images of damaged and undamaged items. After preprocessing the data, a CNN model with convolutional and dense layers was constructed using TensorFlow's Keras API. The model was trained on the training set and evaluated on the test set, achieving remarkable accuracy. The performance was further assessed using metrics like accuracy score, confusion matrix, and classification report, demonstrating the effectiveness of the developed model in accurately detecting damage in images.